pub nonsense

+++ src/internal/root.rs
@@ -292,10 +292,10 @@ impl Root {
             .iter()
             .filter(|squadron| {
                 if squadron.mutables.read().unwrap().ghost {
-                    ((squadron.get_strength(self.config.battle_scalars.avg_duration) as f32)
-                        < (squadron.ideal_strength as f32 * squadron.class.disband_threshold))
+                    (squadron.get_strength(self.config.battle_scalars.avg_duration) as f32)
+                        < (squadron.ideal_strength as f32 * squadron.class.disband_threshold)
                 } else {
-                     0.0 >= (squadron.get_strength(self.config.battle_scalars.avg_duration) as f32)
+                     0 >= squadron.get_strength(self.config.battle_scalars.avg_duration)
                 }
             })
             .cloned()
